# Reduce the number of write operation on H2