--- a/proxy/http2/HTTP2.cc
+++ b/proxy/http2/HTTP2.cc
@@ -805,7 +805,7 @@ float Http2::min_avg_window_update             = 2560.0;
 uint32_t Http2::con_slow_log_threshold         = 0;
 uint32_t Http2::stream_slow_log_threshold      = 0;
 uint32_t Http2::header_table_size_limit        = 65536;
-uint32_t Http2::write_buffer_block_size_index  = 11;
+uint32_t Http2::write_buffer_block_size        = 262144;
 float Http2::write_threshold                   = 0.5;
 
 void
@@ -834,7 +834,7 @@ Http2::init()
   REC_EstablishStaticConfigInt32U(con_slow_log_threshold, "proxy.config.http2.connection.slow.log.threshold");
   REC_EstablishStaticConfigInt32U(stream_slow_log_threshold, "proxy.config.http2.stream.slow.log.threshold");
   REC_EstablishStaticConfigInt32U(header_table_size_limit, "proxy.config.http2.header_table_size_limit");
-  REC_EstablishStaticConfigInt32U(write_buffer_block_size_index, "proxy.config.http2.write_buffer_block_size_index");
+  REC_EstablishStaticConfigInt32U(write_buffer_block_size, "proxy.config.http2.write_buffer_block_size");
   REC_EstablishStaticConfigFloat(write_threshold, "proxy.config.http2.write_threshold");
 
   // If any settings is broken, ATS should not start