--- a/proxy/http2/HTTP2.h
+++ b/proxy/http2/HTTP2.h
@@ -404,7 +404,8 @@ public:
   static uint32_t stream_slow_log_threshold;
   static uint32_t header_table_size_limit;
   static uint32_t write_buffer_block_size;
-  static float write_threshold;
+  static float write_size_threshold;
+  static uint32_t write_time_threshold;
 
   static void init();
 };