--- a/proxy/http2/Http2ClientSession.h
+++ b/proxy/http2/Http2ClientSession.h
@@ -101,7 +101,7 @@ public:
 
   // more methods
   void write_reenable();
-  int64_t xmit(const Http2TxFrame &frame);
+  int64_t xmit(const Http2TxFrame &frame, bool flush = true);
 
   ////////////////////
   // Accessors
@@ -182,6 +182,10 @@ private:
 
   Event *_reenable_event = nullptr;
   int _n_frame_read      = 0;
+
+  MIOBuffer *_send_frame_buffer             = nullptr;
+  IOBufferReader *_send_frame_buffer_reader = nullptr;
+  uint32_t _send_frame_buffer_used_size     = 0;
 
   int64_t read_from_early_data   = 0;
   bool cur_frame_from_early_data = false;