@maskit 2020-10-21
I'm going to make the block size and threshold configurable.