@maskit 2020-10-22
Added two settings and the detail is documented, and made the default buffer size small a little (256KB) because 1MB may be too big for some people.