@maskit 2020-10-26
Addressed Bryan's concern. Time threshold is also added.