@zwoop 2020-10-28
Cherry-picked to v9.0.x branch.